requirements.txt
streamlit==1.24.0
Pillow
requests
python-dotenv
